refactor and support multiple compose environments
