+++ README.md
@@ -28,7 +28,11 @@ uvicorn main:app
 ## Run the API (docker)
 **Build and run the api with docker-compose up:**
 ```bash
+# run a dev environment
 docker-compose --env-file ./config/.env.dev up
+
+# run a 'test' environment, no dev requirements install
+docker-compose --env-file ./config/.env.test up
 ```
 
 You can make requests to the api endpoints such as: http://localhost:8000/players
@@ -36,3 +40,8 @@ You can make requests to the api endpoints such as: http://localhost:8000/player
 You can see the interactive documentation at: http://localhost:8000/docs
 
 You can access the PGAdmin Interface at: http://localhost:80
+
+### Execute Tests within docker-compose (dev environment):
+```bash
+docker-compose exec api sh /usr/test/test-all.sh
+```
