+++ config/.env.test
@@ -1,5 +1,5 @@
 # requirements
-PY_ENV=dev
+PY_ENV=test
 
 # db-pg service
 POSTGRES_USER=postgres
