+++ requirements-dev.txt
@@ -1,0 +1,2 @@
+pytest==7.1.2
+flake8==4.0.1
