+++ src/models/schemas/campaigns.py
