+++ src/models/schemas/characters.py
