+++ src/routes/characters.py
