+++ src/routes/players.py
